Resolve conflict in lib/misc/configuration.xml: keep both <telemetry>
(OTEL config, from this branch) and <admin_api> (from master).

+++ src/engine/core/handler.cpp
@@ -449,7 +449,7 @@ void PlaceObjToInventory(ObjData *object, CharData *ch) {
 				inworld = 1;
 				// Объект готов для проверки. Ищем в мире такой же.
 				world_objects.foreach_with_vnum(GET_OBJ_VNUM(object), [&inworld, tuid, object](const ObjData::shared_ptr &i) {
-					if (i->get_unique_id() == tuid // UID совпадает
+					if (static_cast<unsigned int>(i->get_unique_id()) == tuid // UID совпадает
 						&& i->get_timer() > 0  // Целенький
 						&& object != i.get()) // Не оно же
 					{
